🔨 refactor: extract SelectBox to own file
